html/browsers/browsing-the-web/history-traversal/ScrollRestorationOrder.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>History restoration order test</title>
<meta name="assert" content="http://www.w3.org/TR/2011/WD-html5-20110113/history.html#history-traversal">
<meta name="assert" content="Traversing history should restore scroll position before dispatching popstate">

<style>
  body {
    height: 200vw;
    width: 200vw;
  }
</style>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  'use strict';
  async_test(function(t) {
    window.addEventListener('load', t.step_func(function() {
      assert_array_equals(scrollPosition(), [0, 0]);

      history.pushState('#1', '', '#1');
      window.scrollTo(50, 100);
      assert_array_equals(scrollPosition(), [50, 100]);

      history.pushState('#2', '', '#2');
      window.scrollTo(100, 200);
      assert_array_equals(scrollPosition(), [100, 200]);

      // set scroll position and kick off the verification
      setTimeout(function() {
        history.pushState(null, null, '#done');
        window.scrollTo(555, 555);
        assert_array_equals(scrollPosition(), [555, 555]);
        window.history.back();
      }, 0);
    }));

    window.addEventListener('popstate', t.step_func(function() {
      var key = location.hash;
      switch (key) {
        case '#2':
          assert_equals(history.state, key, 'state is not ' + key);
          assert_array_equals(scrollPosition(), [100, 200], 'scroll is not restored before popstate for ' + key);
          break;
        case '#1':
          assert_equals(history.state, key, 'state is not ' + key);
          assert_array_equals(scrollPosition(), [50, 100], 'scroll is not restored before popstate for ' + key);
          break;
        case '':
          assert_array_equals(scrollPosition(), [0, 0], 'scroll is not restored before popstate for ' + key);
          t.done();
          return;
        default:
          assert_unreached();
          return;
      }

      window.history.back();
    }));

    function scrollPosition() {
      return [window.pageXOffset, window.pageYOffset];
    }

  }, 'Traversing history should restore scroll position before dispatching popstate');
</script>
